refactor github handler
* remove api versioning
* move github handler out of main
