+++ src/main.rs
@@ -1,25 +1,14 @@
 use axum::{
-    async_trait,
     body::Body,
     error_handling::HandleErrorLayer,
-    extract::{FromRequest, Path, RequestParts},
     handler::Handler,
     http::{uri, Request, StatusCode},
-    response::{IntoResponse, Json, Response},
+    response::{IntoResponse, Response},
     routing::get,
     BoxError, Router, Server,
 };
-use serde::de::DeserializeOwned;
-use std::{collections::HashMap, convert::Infallible};
+use std::convert::Infallible;
 use tower::{filter::AsyncFilterLayer, util::AndThenLayer, ServiceBuilder};
-
-#[derive(Debug)]
-enum Version {
-    V1,
-}
-
-#[derive(Debug)]
-struct JsonProxyError(reqwest::Error);
 
 #[tokio::main]
 async fn main() {
@@ -50,7 +39,7 @@ async fn main() {
 
     let app = Router::new()
         .nest("/", meta_routes)
-        .nest("/api/:version", api_routes)
+        .nest("/api", api_routes)
         .layer(middleware_stack)
         .fallback(fallback.into_service());
 
@@ -91,75 +80,3 @@ async fn heartbeat_handler() -> Result<&'static str, StatusCode> {
 async fn fallback() -> impl IntoResponse {
     (StatusCode::NOT_FOUND, "Not Found".to_string())
 }
-
-async fn gh_pages_handler<T>(filename: &str) -> Result<Json<T>, JsonProxyError>
-where
-    T: DeserializeOwned,
-{
-    let url = format!("https://stodevx.github.io/AAO-React-Native/{}", filename).to_string();
-    let resp = request_handler(&url).await?;
-    Ok(resp)
-}
-
-macro_rules! gh_pages_handler {
-    ($name:ident,$filename:literal,$response_type:ty) => {
-        async fn $name(_version: Version) -> Result<Json<$response_type>, JsonProxyError> {
-            let data = gh_pages_handler($filename).await?;
-            Ok(data)
-        }
-    };
-}
-
-macro_rules! gh_pages_handlers {
-    ($([$name:ident, $filename:literal $(, $response_type:ty)?]),+ $(,)?) => {
-        $(
-            gh_pages_handler!($name, $filename $(, $response_type)?);
-        )+
-    };
-}
-
-async fn request_handler<T>(path: &str) -> Result<Json<T>, JsonProxyError>
-where
-    T: DeserializeOwned,
-{
-    let response = reqwest::get(path).await.map_err(JsonProxyError)?;
-
-    response.json().await.map(Json).map_err(JsonProxyError)
-}
-
-impl IntoResponse for JsonProxyError {
-    fn into_response(self) -> axum::response::Response {
-        // `self.0` is a `reqwest::Error`.  Just format it as a string.
-        let body = axum::body::boxed(axum::body::Full::from(self.0.to_string()));
-
-        Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .body(body)
-            .unwrap()
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for Version
-where
-    B: Send,
-{
-    type Rejection = Response;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let params = Path::<HashMap<String, String>>::from_request(req)
-            .await
-            .map_err(IntoResponse::into_response)?;
-
-        let version = params
-            .get("version")
-            .ok_or_else(|| (StatusCode::NOT_FOUND, "Missing Version").into_response())?;
-
-        match version.as_str() {
-            "v1" => Ok(Version::V1),
-            _ => Err((StatusCode::NOT_FOUND, "Unknown Version").into_response()),
-        }
-    }
-}
-
-mod routes;
